+++ libs/remotion/components/src/lib/foregrounds/ImLive.tsx
@@ -10,6 +10,8 @@ type ImLiveProps = {
     fontColor: string;
     arrow: boolean;
     showText: boolean;
+    showUsername: boolean;
+    username?: string;
     thumbnailBorderColor: string;
 };
 
@@ -20,12 +22,21 @@ export const ImLive: React.FC<ImLiveProps> = ({
     thumbnailBorderColor = 'rgb(145, 71, 255)',
     arrow = true,
     showText = true,
+    showUsername = true,
+    username,
 }) => {
     return (
         <div style={{ width: '100%', height: '100%', overflow: 'hidden', fontFamily: 'Inter' }}>
             <div style={{ width: '100%', height: '100%', display: 'flex', flexDirection: 'column', justifyContent: 'flex-end' }}>
                 <div style={{ width: '100%', alignItems: 'flex-end', paddingTop: '4px' }}>
                     {showText && <h1 style={{ color: fontColor, fontSize: '86px', width: '100%', textAlign: 'center', alignItems: 'flex-end', margin: 0 }}>{text}</h1>}
+                </div>
+                <div style={{ width: '100%', alignItems: 'flex-end', paddingTop: '4px' }}>
+                    {showUsername && (
+                        <h1 style={{ color: fontColor, fontSize: showText ? '38px' : '86px', width: '100%', textAlign: 'center', alignItems: 'flex-end', margin: 0 }}>
+                            {username}
+                        </h1>
+                    )}
                 </div>
                 <div style={{ display: 'flex', width: '100%', height: 'min', alignItems: 'bottom', justifyContent: 'flex-end' }}>
                     <div